# Conflicts:
#	sphinx/application/network/concepts/queries/concept-network-query-chat/src/main/java/chat/sphinx/concept_network_query_chat/model/TribeDto.kt

+++ sphinx/application/network/concepts/queries/concept-network-query-chat/src/main/java/chat/sphinx/concept_network_query_chat/model/TribeDto.kt
@@ -23,7 +23,7 @@ data class TribeDto(
     val deleted: Any?,
     val app_url: String?,
     val feed_url: String?,
-    val feed_type: Long?,
+    val feed_type: Int?,
 ) {
 
     var amount: Long? = null
